requirements.txt
requests==2.27.1
python-dotenv==0.19.2
python-telegram-bot==11.1.0
redis~=4.1.4
vk-api==11.9.7